Changed contract command so it can accept JSON arrays if a contract demands an array as input.
This is a temporary solution to do operations such as creating
organizations (which takes an array of members) until we agree on how we
want people to pass array parameters to contract commands.

The way it was working, the command would get the input type from the
ABI json and the type_converter function would treat user input as a
single element of that type.